+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 ## Update Notes:
 - CPR for v14 does not contain all automations that previously existed in V13.
 - Requires CAT version X.Y.Z.
-- Added Generic: damageBonusToOneRoll, passiveAura, hide
+- Added Generic: damageBonusToOneRoll, passiveAura, hide, correctSpellLink
 ## Bug Fixes:
 - Brutal Strike
-- Generic: multiSingleTarget, correctSpellLink
+- Generic: multiSingleTarget
 - Empowered Evocation
 - Song of Defense
 - Greater Divine Intervention
+- Overchannel
 ## Updated Legacy Automations:
 - Channel Divinity (Cleric)
 - Destroy Undead